feat/fe: MemberJoin, MemberLogin Page
