+++ frontend/schduler/src/App.vue
@@ -1,34 +1,6 @@
 <script setup>
 import { RouterLink, RouterView } from 'vue-router'
 import HelloWorld from './components/HelloWorld.vue'
-import { ref } from 'vue';
-import axios from 'axios';
-
-const memberName = ref('');
-const memberEmail = ref('');
-const memberPassword = ref('');
-const memberNickname = ref('');
-
-
-function registerMember() {
-
-  const registerData = {
-    memberName: memberName.value,
-    memberEmail: memberEmail.value,
-    memberPassword: memberPassword.value,
-    memberNickname: memberNickname.value
-  };
-
-  axios.post(`http://localhost:8080/members`, registerData)
-  .then((res) => {
-    console.log(res);
-  })
-  .catch((error) => {
-    console.log(error);
-  })
-
-}
-
 
 </script>
 
@@ -36,36 +8,13 @@ function registerMember() {
   <header>
     <div class="wrapper">
 
-      <div class="register-form-container">
-        <div class="input-item">
-          <span>이름</span>
-          <input type="text" v-model="memberName">
-        </div>
-        <div class="input-item">
-          <span>이메일</span>
-          <input type="text" v-model="memberEmail">
-        </div>
-        <div class="input-item">
-          <span>비밀번호</span>
-          <input type="text" v-model="memberPassword">
-        </div>
-        <div class="input-item">
-          <span>닉네임</span>
-          <input type="text" v-model="memberNickname">
-        </div>
-
-        <div class="footer">
-          <button type="button" @click="registerMember()">
-            회원가입
-          </button>
-        </div>
-      </div>
-
       <HelloWorld msg="You did it!" />
 
       <nav>
         <RouterLink to="/">Home</RouterLink>
         <RouterLink to="/about">About</RouterLink>
+        <RouterLink to="/join">Join</RouterLink>
+        <RouterLink to="/login">Login</RouterLink>
       </nav>
     </div>
   </header>
@@ -74,25 +23,6 @@ function registerMember() {
 </template>
 
 <style scoped>
-
-.register-form-container {
-  display: flex;
-  flex-wrap: wrap;
-  width: 50%;
-
-}
-
-.input-item {
-  display: flex;
-  justify-content: space-between;
-  border: 1px solid blue;
-  width: 100%;
-}
-
-.footer {
-  display: flex;
-  justify-content: space-around;
-}
 
 header {
   line-height: 1.5;
